// AG-SPA/src/app/income/income-filter/income-order-enum.ts
export enum OrderDirection {
    IncomeAscending = 'IncomeAscending',
    IncomeDescending = 'IncomeDescending',
    DateAscending = 'DateAscending',
    DateDescending = 'DateDescending',
    OperatorNameAscending = 'OperatorNameAscending',
    OperatorNameDescending = 'OperatorNameDescending',
    PerformerNameAscending = 'PerformerNameAscending',
    PerformerNameDescending = 'PerformerNameDescending'
}